functions/requirements.txt
firebase_functions~=0.5.0
requests
pandas
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
pytz
